requirements.txt
PyYAML>=5.4
python-dotenv>=0.19
requests>=2.25
pandas>=1.3
SQLAlchemy>=1.4
pytest>=7.0
